refactor structure folders
