+++ src/app/layout.tsx
@@ -1,8 +1,9 @@
-import '@/app/styles/input.css';
+import Navbar from '@/components/Navbar';
+import AuthProvider from '@/components/auth/AuthProvider';
+import '@/styles/input.css';
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
-import AuthProvider from "@/app/components/auth/AuthProvider";
-import Navbar from './components/Navbar'
+
 
 const inter = Inter({ subsets: ['latin'] })
