+++ src/components/Header.tsx
@@ -1,5 +1,3 @@
-
-
 function Header() {
   return (
